App/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
mongoengine==0.28.2
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
pymongo==4.6.0
